Move to decomon.utils common methods with decomon.*.utils
Facts:
- Some methods were missing in decomon.utils (add, relu_, minus,
  substract), and existed in decomon.layers.utils
- Other functions are needed for theses ones (eg. get_bound_grid)
- Some methods were duplicated in both modules, and with other decomon.*.utils
- decomon.layers.utils was already importing from decomon.utils (with a
  *!)

Solution:
- Put needed stuff in decomon.utils (there instead of
  decomon.layers.utils to avoid cyclic import issues)
- Choose the better version of duplicated functions
- Import these functions in decomon.layers.utils if also needed
- Update imports in other modules
