+++ decomon/backward_layers/activations.py
@@ -6,22 +6,21 @@ import tensorflow as tf
 import tensorflow.keras.backend as K
 from tensorflow.keras.layers import Layer
 
-from decomon.backward_layers.utils import V_slope, backward_relu_, backward_softplus_
+from decomon.backward_layers.utils import backward_relu_, backward_softplus_
 from decomon.layers import F_FORWARD, F_HYBRID, F_IBP
 from decomon.layers.core import StaticVariables
-from decomon.layers.utils import (
+from decomon.utils import (
+    V_slope,
+    get_linear_hull_relu,
     get_linear_hull_s_shape,
+    get_linear_hull_sigmoid,
+    get_linear_hull_tanh,
     get_linear_softplus_hull,
     get_lower,
     get_upper,
     sigmoid_prime,
     softsign_prime,
     tanh_prime,
-)
-from decomon.utils import (
-    get_linear_hull_relu,
-    get_linear_hull_sigmoid,
-    get_linear_hull_tanh,
 )
 
 ELU = "elu"
