+++ decomon/metrics/loss.py
@@ -4,12 +4,11 @@ import tensorflow.keras.backend as K
 from tensorflow.keras.layers import Flatten, Input, InputLayer, Lambda, Layer
 
 from decomon.layers.activations import softmax as softmax_
-from decomon.layers.core import DecomonLayer
-from decomon.layers.utils import F_FORWARD, F_HYBRID, F_IBP, get_lower, get_upper
+from decomon.layers.core import F_FORWARD, F_HYBRID, F_IBP, DecomonLayer
 from decomon.metrics.utils import categorical_cross_entropy
 from decomon.models import DecomonModel
 from decomon.models.utils import get_mode
-from decomon.utils import set_mode
+from decomon.utils import get_lower, get_upper, set_mode
 
 
 def get_model(model):
