+++ decomon/models/decomon_sequential.py
@@ -18,10 +18,10 @@ from tensorflow.python.keras.utils.generic_utils import has_arg, to_list
 
 from decomon.backward_layers.backward_layers import get_backward as get_backward_
 from decomon.backward_layers.backward_layers import join
-from decomon.backward_layers.utils import S_slope, V_slope, backward_linear_prod
 from decomon.layers.core import Box, StaticVariables
 from decomon.layers.decomon_layers import to_monotonic
-from decomon.layers.utils import get_lower, get_upper, softmax_to_linear
+from decomon.layers.utils import softmax_to_linear
+from decomon.utils import V_slope, get_lower, get_upper
 
 
 # create static variables for varying convex domain
